+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/error/SoraErrorReason.kt
@@ -16,6 +16,9 @@ enum class SoraErrorReason {
     PEER_CONNECTION_CLOSED,
     TIMEOUT,
 
+    // CA証明書の検証に失敗した場合のエラー
+    CA_CERTIFICATE_VALIDATION_FAILED,
+
     // Sora との接続の警告
     ICE_DISCONNECTED,
     PEER_CONNECTION_DISCONNECTED,